up show ss em o.s

+++ database/migrations/2024_09_04_181250_create_solicitacao_os_table.php
@@ -13,14 +13,16 @@ class CreateSolicitacaoOsTable extends Migration
      */
     public function up()
     {
-        Schema::create('solicitacao_os', function (Blueprint $table) {
-            $table->id(); // ID auto-incremental
-            $table->dateTime('datatime'); // Data e hora
-            $table->unsignedBigInteger('emissor')->nullable(); // Chave estrangeira opcional
-            $table->foreign('emissor')->references('id')->on('funcionarios'); // Relacionamento com a tabela funcionarios
-            $table->string('descricao', 300); // Campo de descrição com limite de 300 caracteres
-            $table->timestamps(); // Campos created_at e updated_at
-        });
+        if (!Schema::hasTable('solicitacao_os')) { // Verifica se a tabela já existe
+            Schema::create('solicitacao_os', function (Blueprint $table) {
+                $table->id(); // ID auto-incremental
+                $table->dateTime('datatime'); // Data e hora
+                $table->unsignedBigInteger('emissor')->nullable(); // Chave estrangeira opcional
+                $table->foreign('emissor')->references('id')->on('funcionarios'); // Relacionamento com a tabela funcionarios
+                $table->string('descricao', 300); // Campo de descrição com limite de 300 caracteres
+                $table->timestamps(); // Campos created_at e updated_at
+            });
+        }
     }
 
     /**
@@ -30,6 +32,9 @@ class CreateSolicitacaoOsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('solicitacao_os');
+        if (Schema::hasTable('solicitacao_os')) { // Verifica antes de excluir
+            Schema::dropIfExists('solicitacao_os');
+        }
     }
 }
+
